// src/feature/authorization/api/registration/registration.queryKeys.ts
export enum RegistrationQueryKeys {
  useValidateNationalId = 'useValidateNationalId',
  useInitAbsher = 'useInitAbsher',
  useInitMobile = 'useInitMobile',
  useValidateAbsher = 'useValidateAbsher',
  useValidateEstablishments = 'useValidateEstablishments',
  useValidateEmail = 'useValidateEmail',
  useValidateMobile = 'useValidateMobile',
  useCreateAccount = 'useCreateAccount',
}